fix nillable element
